requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
pymodbus==3.6.2
python-dotenv==1.0.1
twilio==8.12.0
typing-extensions==4.9.0 